refactor: colocate types and keep a single pattern
