+++ src/app/(list)/listTwoColumns/[id].tsx
@@ -1,10 +1,10 @@
 import OrdersProducts from "@/features/lists/listWithTwoColumns/components/OrdersProducts";
-import { GridType } from "@/features/shared/types/grid";
+import { TGrid } from "@/features/lists/listWithTwoColumns/types/grid";
 import { useLocalSearchParams, useNavigation } from "expo-router";
 import { useLayoutEffect } from "react";
 
 export default function ListTwoColumns() {
-  const params = useLocalSearchParams<{ id: GridType }>();
+  const params = useLocalSearchParams<{ id: TGrid }>();
   const navigation = useNavigation();
 
   useLayoutEffect(() => {
