+++ src/data/types/Fruit.ts
@@ -1,4 +1,4 @@
-export type Fruit = {
+export type TFruit = {
   name: string;
   icon: string;
   id: string;
