+++ src/features/home/data/menus.ts
@@ -1,6 +1,6 @@
 import listMenu from "@/features/lists/data/listMenu";
 import animationMenu from "@/features/animations/data/animationMenu";
-import { TMenu } from "../shared/types/menu";
+import { TMenu } from "@/components/Menu/types";
 
 const menus: TMenu[] = [listMenu, animationMenu];
