+++ src/features/lists/longList/components/LongListFlashList.tsx
@@ -1,4 +1,4 @@
-import { Fruit } from "@/data/types/Fruit";
+import { TFruit } from "@/data/types/Fruit";
 import useFruits from "@/hooks/useFruits";
 import { FlashList } from "@shopify/flash-list";
 import React from "react";
@@ -8,11 +8,11 @@ import FruitItem from "./FruitItem";
 const LongListFlashList = () => {
   const { fruits, addFruit } = useFruits();
 
-  const renderItem = ({ item: { name, icon } }: { item: Fruit }) => (
+  const renderItem = ({ item: { name, icon } }: { item: TFruit }) => (
     <FruitItem name={name} icon={icon} />
   );
 
-  const keyExtractor = ({ id }: Fruit) => id;
+  const keyExtractor = ({ id }: TFruit) => id;
   return (
     <>
       <Button title="add item" onPress={addFruit} />
