+++ src/features/lists/longList/components/LongListFlatList.tsx
@@ -2,12 +2,12 @@ import useFruits from "@/hooks/useFruits";
 import React from "react";
 import { Button, FlatList } from "react-native";
 import FruitItem from "./FruitItem";
-import { Fruit } from "@/data/types/Fruit";
+import { TFruit } from "@/data/types/Fruit";
 
 const LongListFlatList = () => {
   const { fruits, addFruit } = useFruits();
 
-  const renderItem = ({ item: { name, icon } }: { item: Fruit }) => (
+  const renderItem = ({ item: { name, icon } }: { item: TFruit }) => (
     <FruitItem name={name} icon={icon} />
   );
